refactor (auth) compare password by hash, optimize fiolders structure
