+++ src/auth/auth.module.ts
@@ -3,8 +3,8 @@ import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { UsersModule } from '../users/users.module';
 import { AuthService } from './auth.service';
-import { JwtStrategy } from './jwt-strategy.service';
-import { LocalStrategy } from './local-strategy.service';
+import { JwtStrategy } from './passport-strategies/jwt-strategy.service';
+import { LocalStrategy } from './passport-strategies/local-strategy.service';
 import { AuthController } from './auth.controller';
 import { JwtConfigFactory } from '../config/jwt-module-config.factory';
 
@@ -13,10 +13,6 @@ import { JwtConfigFactory } from '../config/jwt-module-config.factory';
     imports: [
         UsersModule,
         PassportModule,
-        // JwtModule.register({
-            // secret: 'super secret key!!',
-            // global: true,
-            // signOptions: { expiresIn: '60s'},
         JwtModule.registerAsync({
             useClass: JwtConfigFactory,
         })
