+++ src/auth/auth.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
+import { isCorrectHash } from 'src/shared/hash';
 import { User } from 'src/users/entities/user.entity';
 import { UsersService } from 'src/users/users.service';
 
@@ -11,13 +12,13 @@ export class AuthService {
     ) { }
 
     auth(user: User) {
-        return { access_token: this.jwtService.sign({ id: user.id }) }
+        return { access_token: this.jwtService.sign({ sub: user.id }) }
         // return { access_token: this.jwtService.signAsync({ sub: user.id, username: user.username }) }
     }
 
     async validatePassword(username: string, password: string) {
         const user = await this.userService.findByUsername(username);
-        if (!user || user.password === password) throw new UnauthorizedException('wrong username or password');
+        if (!user || !isCorrectHash(user.password, password)) throw new UnauthorizedException('wrong username or password');
         delete user.password;
         return user;
     }
