+++ src/auth/passport-strategies/jwt-guard.ts
@@ -2,4 +2,5 @@ import { Injectable } from "@nestjs/common";
 import { AuthGuard } from "@nestjs/passport";
 
 @Injectable()
-export class JwtGuard extends AuthGuard('jwt') {}
+export class JwtGuard extends AuthGuard('jwt') { }
+
