+++ src/users/dto/create-user.dto.ts
@@ -1,3 +1,3 @@
 import { PickType } from "@nestjs/mapped-types";
 import { User } from "../entities/user.entity";
-export class CreateUserDto extends PickType(User, ['email'] as const) { }
+export class CreateUserDto extends PickType(User, ['username', 'password', 'email'] as const) { }
